+++ src/Git_Updater/Traits/GU_Trait.php
@@ -599,6 +599,7 @@ trait GU_Trait {
 		$header['ci_job']         = false;
 		$header['release_asset']  = false;
 		$header['primary_branch'] = false;
+		$header['id']             = null;
 
 		if ( ! empty( $header['host'] ) ) {
 			if ( 'GitHub' === $header_parts[0] && ! str_contains( $header['host'], 'github.com' ) ) {
@@ -632,6 +633,9 @@ trait GU_Trait {
 		}
 		$header['release_asset']  = ! $header['release_asset'] && ! empty( $headers['ReleaseAsset'] ) ? true === (bool) $headers['ReleaseAsset'] : $header['release_asset'];
 		$header['primary_branch'] = ! $header['primary_branch'] && ! empty( $headers['PrimaryBranch'] ) ? $headers['PrimaryBranch'] : 'master';
+
+		$header['id'] = ! empty( $headers['PluginID'] ) ? $headers['PluginID'] : '';
+		$header['id'] = ! empty( $headers['ThemeID'] ) ? $headers['ThemeID'] : $header['id'];
 
 		return $header;
 	}